requirements.txt
crewai>=0.28.0
google-generativeai>=0.5.0
sentence-transformers>=2.2.2
arxiv>=2.1.0
pymupdf>=1.23.0
endee>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
transformers>=4.51.0
torch>=2.0.0
